# [js/web] JSEP node assignment optimization