--- a/js/web/lib/wasm/jsep/webgpu/ops/gather.ts
+++ b/js/web/lib/wasm/jsep/webgpu/ops/gather.ts
@@ -41,30 +41,21 @@ const createGatherProgramInfo =
       const gatheredBatchElements = N * block * elementSize;
       const axisDimLimit = inputShape[axis];
 
-      const inputIndices: number[] = [];
-      const inputIndicesArray = indicesElementSize === 2
-        ? inputs[1].getBigInt64Array()
-        : inputs[1].getInt32Array();
-      for (const i of inputIndicesArray[Symbol.iterator]()) {
-        let value = Number(i);
-        if (value < 0) {
-          value += axisDimLimit;
-        }
-        inputIndices.push(value);
-      }
-
       const inputSize = ShapeUtil.size(inputShape) * elementSize;
       const outputSize = ShapeUtil.size(outputShape) * elementSize;
 
       const totalGathers = M * N;
+      // int64 indices would be treated as little endian i32 with assumption they fall in i32 limits
+      // That assumption is safe as it's not possible to allocate >2gb buffer for input tensor
+      // Input data will be treated as u32 or two u32 for 8-byte tensors
       const getShaderSource = (shaderHelper: ShaderHelper) => `
   const N: u32 = ${N};
   const elementSize: u32 = ${elementSize};
   const indicesElementSize: u32 = ${indicesElementSize};
-  const inputIndices = array<u32, N>(${inputIndices.map(i => `${i}u`).join(',')});
 
   @group(0) @binding(0) var<storage, read> input : array<u32>;
-  @group(0) @binding(1) var<storage, read_write> output: array<u32>;
+  @group(0) @binding(1) var<storage, read> inputIndices : array<i32>;
+  @group(0) @binding(2) var<storage, read_write> output: array<u32>;
 
   ${shaderHelper.mainStart()}
     let batch: u32 = global_idx / N;
@@ -72,9 +63,12 @@ const createGatherProgramInfo =
 
     let srcOffsetBatch: u32 = batch * ${dataBatchElements};
     let dstOffsetBatch: u32 = batch * ${gatheredBatchElements};
-    var idx = inputIndices[i];
+    var idx = inputIndices[i * indicesElementSize];
+    if (idx < 0) {
+        idx = idx + ${axisDimLimit};
+    }
 
-    let srcOffset = srcOffsetBatch + idx * ${blockSize};
+    let srcOffset = srcOffsetBatch + u32(idx) * ${blockSize};
     let dstOffset = dstOffsetBatch + i * ${blockSize};
     if (srcOffset >= ${inputSize}) {
         return;
@@ -105,13 +99,9 @@ export const gather = (context: ComputeContext, attributes: GatherAttributes): v
 
   const metadata = {
     name: 'Gather',
-    inputTypes: [GpuDataType.default],
-    cacheHint: attributes.cacheKey + inputs[0].dataType.toString(10) + inputs[1].dataType.toString(10) +
-    inputs[1].dims.join(','),
+    inputTypes: [GpuDataType.default, GpuDataType.default],
+    cacheHint: attributes.cacheKey,
   };
 
-  context.compute(
-    createGatherProgramInfo(metadata, context.inputs, attributes),
-    { inputs: [inputs[0]] },
-  );
+  context.compute(createGatherProgramInfo(metadata, context.inputs, attributes));
 };