--- a/onnxruntime/core/providers/js/js_data_types.cc
+++ b/onnxruntime/core/providers/js/js_data_types.cc
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+
 #include "core/providers/cpu/tensor/shape_op.h"
 
 namespace onnxruntime {
@@ -14,5 +15,5 @@ using SupportedTypes =
 std::vector<MLDataType> JsepSupportedDataTypes() {
   return BuildKernelDefConstraintsFromTypeList<SupportedTypes>();
 }
-}
-}
+}  // namespace js
+}  // namespace onnxruntime