--- a/onnxruntime/core/providers/js/js_data_types.cc
+++ b/onnxruntime/core/providers/js/js_data_types.cc
@@ -12,7 +12,9 @@ using SupportedTypes =
         int32_t,
         uint32_t>;
 
-const std::vector<MLDataType> JsepSupportedDataTypes = BuildKernelDefConstraintsFromTypeList<SupportedTypes>();
-
+const std::vector<MLDataType>& JsepSupportedDataTypes() {
+  static const std::vector<MLDataType> supportedDataTypes = BuildKernelDefConstraintsFromTypeList<SupportedTypes>();
+  return supportedDataTypes;
+}
 }  // namespace js
 }  // namespace onnxruntime