--- a/onnxruntime/core/providers/js/js_data_types.h
+++ b/onnxruntime/core/providers/js/js_data_types.h
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#include "core/framework/data_types.h"
+
+namespace onnxruntime {
+namespace js {
+  std::vector<MLDataType> JsepSupportedDataTypes();
+}
+}