--- a/onnxruntime/core/providers/js/operators/gather.cc
+++ b/onnxruntime/core/providers/js/operators/gather.cc
@@ -8,15 +8,6 @@
 namespace onnxruntime {
 namespace js {
 
-using AllSupportedSize =
-    TypeList<
-        float,
-        double,
-        int64_t,
-        uint64_t,
-        int32_t,
-        uint32_t>;
-
 ONNX_OPERATOR_VERSIONED_KERNEL_EX(
     Gather,
     kOnnxDomain,
@@ -24,9 +15,8 @@ ONNX_OPERATOR_VERSIONED_KERNEL_EX(
     10,
     kJsExecutionProvider,
     (*KernelDefBuilder::Create())
-        .TypeConstraint("T", JsepSupportedDataTypes)
-        .TypeConstraint("Tind", BuildKernelDefConstraintsFromTypeList<TypeList<int32_t, int64_t>>())
-        .InputMemoryType(OrtMemTypeCPU, 1),
+        .TypeConstraint("T", JsepSupportedDataTypes())
+        .TypeConstraint("Tind", BuildKernelDefConstraintsFromTypeList<TypeList<int32_t, int64_t>>()),
     Gather);
 
 ONNX_OPERATOR_VERSIONED_KERNEL_EX(
@@ -36,9 +26,8 @@ ONNX_OPERATOR_VERSIONED_KERNEL_EX(
     12,
     kJsExecutionProvider,
     (*KernelDefBuilder::Create())
-        .TypeConstraint("T", JsepSupportedDataTypes)
-        .TypeConstraint("Tind", BuildKernelDefConstraintsFromTypeList<TypeList<int32_t, int64_t>>())
-        .InputMemoryType(OrtMemTypeCPU, 1),
+        .TypeConstraint("T", JsepSupportedDataTypes())
+        .TypeConstraint("Tind", BuildKernelDefConstraintsFromTypeList<TypeList<int32_t, int64_t>>()),
     Gather);
 
 ONNX_OPERATOR_KERNEL_EX(
@@ -47,9 +36,8 @@ ONNX_OPERATOR_KERNEL_EX(
     13,
     kJsExecutionProvider,
     (*KernelDefBuilder::Create())
-        .TypeConstraint("T", JsepSupportedDataTypes)
-        .TypeConstraint("Tind", BuildKernelDefConstraintsFromTypeList<TypeList<int32_t, int64_t>>())
-        .InputMemoryType(OrtMemTypeCPU, 1),
+        .TypeConstraint("T", JsepSupportedDataTypes())
+        .TypeConstraint("Tind", BuildKernelDefConstraintsFromTypeList<TypeList<int32_t, int64_t>>()),
     Gather);
 
 }  // namespace js