--- a/onnxruntime/core/providers/js/operators/reshape.cc
+++ b/onnxruntime/core/providers/js/operators/reshape.cc
@@ -13,7 +13,7 @@ ONNX_OPERATOR_KERNEL_EX(
     14,
     kJsExecutionProvider,
     (*KernelDefBuilder::Create())
-        .TypeConstraint("T", JsepSupportedDataTypes)
+        .TypeConstraint("T", JsepSupportedDataTypes())
         .TypeConstraint("shape", DataTypeImpl::GetTensorType<int64_t>())
         .Alias(0, 0)
         .InputMemoryType(OrtMemTypeCPU, 1),
@@ -25,7 +25,7 @@ ONNX_OPERATOR_VERSIONED_KERNEL_EX(
     13, 13,
     kJsExecutionProvider,
     (*KernelDefBuilder::Create())
-        .TypeConstraint("T", JsepSupportedDataTypes)
+        .TypeConstraint("T", JsepSupportedDataTypes())
         .TypeConstraint("shape", DataTypeImpl::GetTensorType<int64_t>())
         .Alias(0, 0)
         .InputMemoryType(OrtMemTypeCPU, 1),
@@ -37,7 +37,7 @@ ONNX_OPERATOR_VERSIONED_KERNEL_EX(
     5, 12,
     kJsExecutionProvider,
     (*KernelDefBuilder::Create())
-        .TypeConstraint("T", JsepSupportedDataTypes)
+        .TypeConstraint("T", JsepSupportedDataTypes())
         .TypeConstraint("shape", DataTypeImpl::GetTensorType<int64_t>())
         .Alias(0, 0)
         .InputMemoryType(OrtMemTypeCPU, 1),