--- a/onnxruntime/core/providers/js/operators/unsqueeze.cc
+++ b/onnxruntime/core/providers/js/operators/unsqueeze.cc
@@ -13,7 +13,7 @@ ONNX_OPERATOR_KERNEL_EX(
     13,
     kJsExecutionProvider,
     (*KernelDefBuilder::Create())
-        .TypeConstraint("T", JsepSupportedDataTypes())
+        .TypeConstraint("T", JsepSupportedDataTypes)
         .TypeConstraint("axes", DataTypeImpl::GetTensorType<int64_t>())
         .Alias(0, 0)
         .InputMemoryType(OrtMemTypeCPU, 1),
@@ -25,7 +25,7 @@ ONNX_OPERATOR_VERSIONED_KERNEL_EX(
     11, 12,
     kJsExecutionProvider,
     (*KernelDefBuilder::Create())
-        .TypeConstraint("T", JsepSupportedDataTypes())
+        .TypeConstraint("T", JsepSupportedDataTypes)
         .Alias(0, 0),
     Unsqueeze);
 
@@ -35,7 +35,7 @@ ONNX_OPERATOR_VERSIONED_KERNEL_EX(
     1, 10,
     kJsExecutionProvider,
     (*KernelDefBuilder::Create())
-        .TypeConstraint("T", JsepSupportedDataTypes())
+        .TypeConstraint("T", JsepSupportedDataTypes)
         .Alias(0, 0),
     Unsqueeze);